create jwt encode interface (#31)
