+++ twimi-app/Cargo.toml
@@ -17,6 +17,7 @@ actix-web = "4.2.1"
 anyhow = "1.0.66"
 argon2 = { version = "0.4.1", features = ["std"] }
 async-trait = "0.1.58"
+jsonwebtoken = "8.1.1"
 shaku = "0.6.1"
 shaku_actix = "0.2.0"
 serde = { version = "1.0.147", features = ["derive"] }
