+++ twimi-app/src/lib.rs
@@ -3,10 +3,10 @@ pub mod server;
 
 use shaku::module;
 
-use infrastructure::services::{Argon2PasswordHasher, Argon2PasswordVerifier};
+use infrastructure::services::{Argon2PasswordHasher, Argon2PasswordVerifier, JwtEncoderImpl};
 use infrastructure::{
     repositories::{health_check::HealthCheckRepositoryImpl, users::UsersRepositoryImpl},
-    MySqlDatabase,
+    JwtSecret, MySqlDatabase,
 };
 use twimi_core::usecases::{
     health_check::HealthCheckUseCase, login::LoginUseCase, signup::SignUpUseCase,
@@ -16,6 +16,7 @@ module! {
     pub AppModule {
         components = [
             MySqlDatabase,
+            JwtSecret,
             HealthCheckRepositoryImpl,
             UsersRepositoryImpl,
             HealthCheckUseCase,
@@ -23,6 +24,7 @@ module! {
             LoginUseCase,
             Argon2PasswordHasher,
             Argon2PasswordVerifier,
+            JwtEncoderImpl,
         ],
         providers = []
     }
