+++ twimi-core/Cargo.toml
@@ -10,9 +10,6 @@ path = "src/lib.rs"
 [dependencies]
 anyhow = "1.0.66"
 async-trait = "0.1.58"
-base64 = "0.13.1"
-jsonwebtoken = "8.1.1"
-serde = { version = "1.0.147", features = ["derive"] }
 shaku = "0.6.1"
 thiserror = "1.0.37"
 time = { version = "0.3.17", features = ["serde"] }
